+++ internal/model/datasource.go
@@ -12,13 +12,19 @@ type Datasource struct {
 type DatasourceSource struct {
 	Fake       *FakeDatasource       `json:"fake,omitempty"`
 	Prometheus *PrometheusDatasource `json:"prometheus,omitempty"`
+	Graphite   *GraphiteDatasource   `json:"graphite,omitempty"`
 }
 
 // FakeDatasource is the fake datasource.
 type FakeDatasource struct{}
 
-// PrometheusDatasource is the Prometheus kind datasource..
+// PrometheusDatasource is the Prometheus kind datasource.
 type PrometheusDatasource struct {
+	Address string `json:"address,omitempty"`
+}
+
+// GraphiteDatasource is the Graphite kind datasource.
+type GraphiteDatasource struct {
 	Address string `json:"address,omitempty"`
 }
 
@@ -33,6 +39,8 @@ func (d Datasource) Validate() error {
 	switch {
 	case d.Prometheus != nil:
 		err = d.Prometheus.validate()
+	case d.Graphite != nil:
+		err = d.Graphite.validate()
 	case d.Fake != nil:
 	default:
 		err = fmt.Errorf("declared datasource %s can't be empty", d.ID)
@@ -47,6 +55,14 @@ func (d Datasource) Validate() error {
 func (p PrometheusDatasource) validate() error {
 	if p.Address == "" {
 		return fmt.Errorf("prometheus address can't be empty")
+	}
+
+	return nil
+}
+
+func (g GraphiteDatasource) validate() error {
+	if g.Address == "" {
+		return fmt.Errorf("Graphite API address can't be empty")
 	}
 
 	return nil